:sparkles: 修改参数限制，优化字典

+++ src/server/article-cat/article-cat.controller.ts
@@ -84,6 +84,13 @@ export class ArticleCatController {
   @UseGuards(JwtAuthGuard)
   @ApiOperation({ summary: '删除' })
   async delete(@CurUser() curUser, @Body() baseFindByIdDto: BaseFindByIdDto): Promise<any> {
+    let { id } = baseFindByIdDto;
+    let isExistId = await this.articleCatService.isExistId(id);
+
+    if (!isExistId) {
+      throw new BadRequestException(`数据 id：${id} 不存在！`);
+    }
+
     return await this.articleCatService.deleteById(baseFindByIdDto);
   }
 }
